Add initial code drop
Early (mostly?) working version of `check_illiad_emails`
plugin.

- Check for "Pending Emails" count greater than configured
  `WARNING` and `CRITICAL` thresholds
  - usable default values provided if not specified

- Check for "Pending Emails" with an age greater than
  configured `WARNING` and `CRITICAL` thresholds
  - usable default values provided if not specified
  - has precedence over pending emails count check

Further refinement, refactoring is needed, but overall this
is a functional form for a first v0.1.0 release. This is
provided that further testing proves no obvious defects
are present.

Documentation updates have not been applied yet; the README
is little more than a placeholder, as is the Go Doc file.

refs GH-1